feat: add Spin SDK for Rust
This commit adds a Rust SDK for Spin, which initially facilitates
building HTTP components.

The HTTP SDK uses objects from the `http`
crate, as opposed to directly exposing the generated objects from the
Spin HTTP interface. This has the advantage of making it significantly
easier to work with requests and responses.

This also exposes the WASI experimental HTTP crate as part of the SDK,
which uses the same request and response objects.

Signed-off-by: Radu Matei <radu.matei@fermyon.com>

Co-authored-by: Brian Hardock <brian.hardock@fermyon.com>
Co-authored-by: Lann Martin <lann.martin@fermyon.com>
Co-authored-by: Radu Matei <radu.matei@fermyon.com>