chore!: update `wasmtime` and `wit-bindgen` dependencies
`wasmtime` 0.35.3 is, as of this writing, the most recent version of `wasmtime`
that supports module linking.  Later versions have removed module linking in
favor of the component model, but the component model work is still in progress,
so we can't upgrade to the very latest until that work is complete.

BREAKING CHANGE: This also updates `wit-bindgen` to a snapshot of the main
branch as of June 13th.  Known breaking changes from that update include:

- breaking ABI changes -- you'll need to rebuild modules using a matching
  `wit-bindgen`

- "function" has been renamed to "func" in WIT files

- pull- and push-buffer support has been removed

Due to the latter, I've removed `crates/object-store` from this repo
temporarily.  We can add it back in once stream support has been implemented in
`wit-bindgen`.

Note that I've regenerated the Go SDK using the new `wit-bindgen`.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>